Renamed view pages
