+++ src/views/Banner/BannerView.tsx
